pyproject.toml
[project]
name = "projeto-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=7.2.0",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.12",
    "pyarrow>=24.0.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "streamlit>=1.56.0",
]
